frontend/src/component/AIModule/JobRecommend/new_cvv_screen/requirements.txt
flask
spacy
pymupdf
python-docx
Flask
PyMuPDF
python-docx
spacy
